moved firebase addData function to client side
